requirements.txt
flask==2.3.2
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-adapter-teams==4.15.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0